DSL hardening (Issue #54): - Fix boolean short-circuit semantics: treat None as False, preserve operand values for AND/OR - Introduce stable exception module (dsl_exceptions) to avoid class identity issues on reload - Clarify package discovery in pyproject.toml to include only 'nav_insights' (fix editable install error)
